% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dst.R
\name{read_dst}
\alias{read_dst}
\title{Read profile data from StarOddi}
\usage{
read_dst(fil)
}
\arguments{
\item{fil}{Name of the file to read}
}
\value{
A tibble with metadata stored in attributes
}
\description{
Read profile data from StarOddi
}
\examples{
library(staroddi)
dst <- read_dst(system.file("demos/1M9380.DAT", package="staroddi"))
dst
attributes(dst)$header
}
