% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dst_header.R
\name{read_dst_header}
\alias{read_dst_header}
\title{Read header data, normally from .DAT file}
\usage{
read_dst_header(fil, long = TRUE)
}
\arguments{
\item{fil}{file name, including path}

\item{long}{boolean(default TRUE). Use FALSE better intend to read in lots of files}
}
\value{
a tibble with header data and detail file information
}
\description{
The data header in .DAT file containing a detailed description
of how the *.DAT file is configured. This feature is used by SeaStar when graphically
}
\examples{
library(staroddi)
header <- read_dst_header(system.file("demos/1M9380.DAT", package = "staroddi"))
header
}
